+++ R/evaluate.R
@@ -1,3 +1,97 @@
+#' Function to test information criteria. Not exported and used by internal
+#' function `kchoose`.
+#' 
+#' @param cl 
+#' Output from `clustra` function.
+#' @param data 
+#' A valid data set for `clustra`.
+#' @param fn
+#' Character file name for output.
+#' 
+#' @return
+#' Numerical value of computed AIC. Also writes a line of computed information 
+#' criteria to `fn` file for each k.
+#' 
+ic_fun = function(cl, data, fn) {
+  group = NULL # for data.table R CMD check
+  
+  k = length(cl$tps)
+  N = sum(cl$counts_df)
+  
+  aic = unlist(lapply(cl$tps, function(x) x$aic))
+  deviance = sum(unlist(lapply(cl$tps, function(x) x$deviance)))
+  ssq = sum(unlist(lapply(1:k, m = cl$tps, function(i, m) 
+     sum((predict(m[[i]], data[group == i]) - data[group == i]$response)^2))))
+  df.null = unlist(lapply(cl$tps, function(x) x$df.null))
+  df.residual = unlist(lapply(cl$tps, function(x) x$df.residual))
+  edf = unlist(lapply(cl$tps, function(x) sum(x$edf)))
+  AIC = ssq/N + 2*sum(edf) 
+  
+  ## (7.30) in ESL (Hastie & Tibshirani)
+  sig2 = ssq/(N - sum(edf))
+  ## agrees with weighted result from gam/bam of mgcv
+  ##sig2w = sum(cl$counts_df * unlist(lapply(cl$tps, function(x) sum(x$sig2))))/N
+  AICht = ssq/N + 2*(sum(edf)/N)*sig2
+  BIC = ssq/N + log(sum(cl$counts_df))*sum(edf)
+  BICht = (N/sig2)*(ssq/N + log(N)*(sum(edf)/N)*sig2) ## (7.36) in ESL (Hastie & Tibshirani)
+  BICrams = ssq/sig2 + log(N)*(sum(edf))
+  
+  ## Oddly, the standard AIC and BIC seem to work well and the adjusted "ht" 
+  ## and "rams" do not!? Is the standard correct??
+  fmt = "%d %7.0f %6.0f %6.0f %6.0f %6.0f %6.0f %10.0f %10.0f"
+  outline = sprintf(fmt, k, sum(aic), AIC, AICht, BIC, BICht, BICrams, deviance,
+                    ssq)
+  cat(outline, "\n", file = fn, append = TRUE)
+  
+  AIC
+}
+
+#' A test function to evaluate information criteria for several k values. Not
+#' exported and only for debugging internal use.
+#' 
+#' @param K
+#' Integer vector of k values to try.
+#' @param var
+#' A numerical value of noise variance in generated data.
+#' @param maxdf 
+#' Fitting parameters. See \code{\link{trajectories}}.
+#' @param mc 
+#' Number of cores to use. Increase up to largest k, or number of cores 
+#' available, whichever is less. (On hyperthreaded cores, up to 2x number of 
+#' cores.)
+#' @param fn 
+#' Character file name for output.
+#' 
+kchoose = function(K, var = 5, maxdf = 10, mc = 1, fn = "ic.txt") {
+  library(clustra)
+  mc = 4 # If running on a unix or a Mac platform, increase up to 2x # cores
+  init = "distant" # maximize median absolute error for initial traj selection
+  set.seed(12345)
+  data = gen_traj_data(n_id = c( 400, 400, 800, 800, 1600), 
+                       types = c(1, 2, 3, 2, 1),
+                       intercepts = c(100, 80, 120, 60, 70), m_obs = 25, 
+                       s_range = c(-365, -14), e_range = c(0.5*365, 2*365),
+                       noise = c(0, var))
+  cat("k   aic     AIC  AICht   BIC   BICht   BICrams  deviance ssq\n", file = fn)
+  plot_sample(data, layout = c(3,3))
+    
+  for(k in K) {
+    set.seed(1234737)
+    cl = clustra(data, k, starts = init, maxdf = maxdf, conv = c(10, 0), 
+                 mccores = mc, verbose = TRUE)
+    plot_smooths(data, cl$tps)
+    ic_fun(cl, data, fn = fn)
+  }
+  res = data.table::fread(fn)
+  
+  opar = par(mfrow = c(2, 2))
+  on.exit(par(opar))
+  plot(AIC ~ k, data = res, type = "l")
+  plot(BIC ~ k, data = res, type = "l")
+  plot(deviance ~ k, data = res, type = "l")
+  plot(ssq ~ k, data = res, type = "l")
+}
+
 #' allpair_RandIndex: helper for replicated cluster comparison
 #' 
 #' Runs \code{\link[MixSim]{RandIndex}} for all pairs of cluster results in its
@@ -45,106 +139,6 @@ allpair_RandIndex = function(results) {
   ri_df
 }
 
-#' Matrix plot of Rand Index comparison of replicated clusters
-#' 
-#' @param rand_pairs
-#' A data frame result of \code{\link{allpair_RandIndex}}
-#' @param name 
-#' Character string file name for pdf plot. If omitted or NULL, plot will 
-#' render to current graphics device.
-#'
-#' @return
-#' Invisible. Full path name of file with plot.
-#'
-#' @references
-#' Wei-chen Chen, George Ostrouchov, David Pugmire, Prabhat, and Michael Wehner.
-#' 2013. A Parallel EM Algorithm for Model-Based Clustering Applied to the
-#' Exploration of Large Spatio-Temporal Data. Technometrics, 55:4, 513-523.
-#'
-#' Sorts replicates within cluster K
-#' Assumes K starts from 2
-#' 
-#' @author
-#' Wei-Chen Chen and George Ostrouchov
-#' 
-#' @importFrom 
-#' grDevices colorRampPalette dev.off pdf
-#' @importFrom 
-#' graphics abline axis box image layout par
-#' @export
-rand_plot = function(rand_pairs, name = NULL) {
-  K.vec = unique(unlist(rand_pairs[, c("i.K", "j.K")]))
-  K.max = max(K.vec)
-  K.len = length(K.vec)
-  R.vec = unique(unlist(rand_pairs[, c("i.R", "j.R")]))
-  R.len = length(R.vec)
-  R.max = max(R.vec)
-  n.col = K.len*R.len
-
-  ## axes
-  x = 0:((K.max - 1) * R.max)
-  y = x
-  
-  ## fill square array
-  z = array(NA, c(max(x), max(x)))
-  diag(z) = 1
-  var.z = rand_pairs$adjRand
-  for(i in 1:nrow(rand_pairs)){ # allows non-sequential K values
-    Kbase = function(val) (which(K.vec == val) - 1)*R.max
-    z.i = Kbase(rand_pairs$i.K[i]) + rand_pairs$i.R[i]
-    z.j = Kbase(rand_pairs$j.K[i]) + rand_pairs$j.R[i]
-    z[z.i, z.j] = z[z.j, z.i] = var.z[i]
-  }
-  lim = range(x)
-  zlim = range(c(rand_pairs$adjRand, 1))
-  ## made by colors = brewer.pal(9, "YlOrRd") # removed dependency
-  colors = c("#FFFFCC", "#FFEDA0", "#FED976", "#FEB24C", "#FD8D3C", "#FC4E2A",
-             "#E31A1C", "#BD0026", "#800026")
-  var.col = colorRampPalette(colors)(n.col)
-  
-  ## Reorder within K for vis effect
-  for(i in 1:K.len) { 
-    iv = (i - 1) * R.max + (1:R.max)
-    for(j in R.len:1) { # using later rows for breaking ties in earlier rows
-      ic = (i - 1) * R.max + j
-      id.z = order(z[iv, ic], decreasing = TRUE)
-      tmp.z = z[iv, ]
-      z[iv, ] = tmp.z[id.z,]
-      tmp.z = z[, iv]
-      z[, iv] = tmp.z[, id.z]
-    }
-  }
-
-  bg = "transparent"
-  fg = "black"
-  wd = 6.5
-  ht = 6
-  if(is.character(name)) pdf(name, width = wd, height = ht, bg = bg)
-    oldpar <- par(no.readonly = TRUE)
-    on.exit(par(oldpar)) # restore old par when exit
-    par(bg = bg, fg = fg, col = fg, col.axis = fg, col.lab = fg, col.main = fg,
-        col.sub = fg, fig = c(0, ht/wd, 0, 1),
-        mar = par()$mar + c(0, 1, 0, -0.5))
-    image(x, y, z, zlim, lim, rev(lim), col = var.col, axes = FALSE,
-          xlab = "Number of Clusters", ylab = "Number of Clusters",
-          main = "Adjusted Rand Index")
-    abline(h = (1:K.len) * R.max, lty = 1, col = "black")
-    abline(v = (1:K.len) * R.max, lty = 1, col = "black")
-    axis(1, at = ((0:(K.len - 1)) + 0.5) * R.max, labels = K.vec)
-    axis(2, at = ((0:(K.len - 1)) + 0.5) * R.max, labels = K.vec)
-    box()
-
-    par(bg = bg, fg = fg, col = fg, col.axis = fg, col.lab = fg, col.main = fg,
-        col.sub = fg, mar = c(5, 0, 4, 2), fig = c(ht/wd, 1, 0, .9), new = TRUE)
-    z.label = seq(zlim[1], zlim[2], length = n.col)
-    image(1, z.label, matrix(z.label, nrow = 1), zlim, c(1, 1), zlim,
-          col = var.col, axes = FALSE, xlab = "", ylab = "")
-    axis(4)
-    box()
-  if(is.character(name)) dev.off()
-  layout(1)
-  invisible(name)
-}
 
 #' clustra_sil: Prepare silhouette plot data for several k or for a previous 
 #' clustra run
@@ -163,6 +157,8 @@ rand_plot = function(rand_pairs, name = NULL) {
 #' @param kv
 #' Vector of `clustra` `k` values to run. If `data` is the output from a 
 #' completed `clustra` run, leave `kv` as NULL.
+#' @param starts
+#' See \code{\link{clustra}}.
 #' @param mccores
 #' See \code{\link{trajectories}}.
 #' @param maxdf
@@ -189,9 +185,9 @@ rand_plot = function(rand_pairs, name = NULL) {
 #' list with a single element for a single silhouette plot.
 #' 
 #' @export
-clustra_sil = function(data, kv = NULL, mccores = 1, maxdf = 30, conv = c(10, 0),
-                       save = FALSE, verbose = FALSE) {
-
+clustra_sil = function(data, kv = NULL, starts = "random", mccores = 1, 
+                       maxdf = 30, conv = c(10, 0), save = FALSE, 
+                       verbose = FALSE) {
   sil = function(x) {
     ord = order(x)
     ck = ord[1]
@@ -228,8 +224,8 @@ clustra_sil = function(data, kv = NULL, mccores = 1, maxdf = 30, conv = c(10, 0)
     a_0 = deltime()
 
     if(is.data.frame(data)) {
-      f = clustra(data, kj, mccores = mccores, maxdf = maxdf, conv = conv,
-                verbose = verbose)
+      f = clustra(data, kj, starts = starts, mccores = mccores, maxdf = maxdf,
+                  conv = conv, verbose = verbose)
     } else {
       f = list(loss = data$loss)
     }
@@ -269,7 +265,7 @@ clustra_sil = function(data, kv = NULL, mccores = 1, maxdf = 30, conv = c(10, 0)
 #' See return of {\code{\link{trajectories}}}.
 #' 
 traj_rep = function(group, data, k, maxdf, conv) {
-  id = ..group = NULL
+  id = ..group = NULL # for data.table R CMD check
   data[, group:=..group[id]] # expand group to all data
   trajectories(data, k, group, maxdf, conv, 1, verbose = FALSE)
 }
@@ -286,6 +282,8 @@ traj_rep = function(group, data, k, maxdf, conv) {
 #' The data (see \code{\link{clustra}} description).
 #' @param k
 #' Vector of k values to try.
+#' @param starts
+#' See \code{\link{clustra}}.
 #' @param mccores
 #' Number of cores for replicate parallelism via mclapply.
 #' @param replicates
@@ -302,7 +300,7 @@ traj_rep = function(group, data, k, maxdf, conv) {
 #' @return See \code{\link{allpair_RandIndex}}.
 #' 
 #' @export
-clustra_rand = function(data, k, mccores, replicates = 10, maxdf = 30,
+clustra_rand = function(data, k, starts, mccores, replicates = 10, maxdf = 30,
                         conv = c(10, 0), save = FALSE, verbose = FALSE) {
   id = .GRP = ..group = NULL # for data.table R CMD check
   results = vector("list", replicates*length(k))
@@ -322,7 +320,9 @@ clustra_rand = function(data, k, mccores, replicates = 10, maxdf = 30,
     kj = k[j]
     
     ## Set starting groups outside parallel section to guarantee reproducibility
-    grp = lapply(rep(kj, replicates), sample.int, size = n_id, replace = TRUE)
+#    grp = lapply(rep(kj, replicates), sample.int, size = n_id, replace = TRUE)
+    grp = lapply(rep(kj, replicates), start_groups, data = data, starts = starts,
+                 maxdf = maxdf, conv = conv, mccores = 1, verbose = FALSE)
     f = parallel::mclapply(grp, traj_rep, data = data, k = kj, maxdf = maxdf,
                            conv = conv, mc.cores = mccores)
     fer = lapply(f, function(f, maxdf, conv)