+++ R/generate.R
@@ -7,36 +7,39 @@
 #' @param type
 #' Response type, 1 is constant, 2 is a sin curve portion, and 3 is a sigmoid
 #' portion.
+#' @param intercept 
+#' Used to set response at `smin` time value (not 0) and shift all responses 
+#' accordingly.
 #' @param start
 #' Negative integer giving time of first observation.
 #' @param end
 #' Positive integer giving time of last observation.
 #' @param smin
-#' The smallest possible `start` value among all `id`s. Currently not used.
+#' The smallest possible `start` value among all `id`s. Used to align with 
+#' intercept and then dropped.
 #' @param emax
 #' The largest possible `end` value among all `id`s. Used to scale sin and
 #' sigmoid support.
-#' @param reference
-#' A response value for constant response. Also used to scale sin and sigmoid
-#' responses.
 #' @param noise
 #' Standard deviation of zero mean Gaussian noise added to response functions.
 #' @return 
 #' An `n_obs` by 4 matrix with columns `id`, `time`, `response`, `true_group`.
 #' 
-oneid = function(id, n_obs, type, start, end, smin, emax, reference, noise) {
+oneid = function(id, n_obs, type, intercept, start, end, smin, emax, noise) {
   id = rep(id, n_obs)
   true_group = rep(type, n_obs)
-  time = c(start, sort(floor(runif(n_obs - 3, min = start, max = end))), end)
+  time = c(smin, start, sort(floor(runif(n_obs - 3, min = start, max = end))), end)
   time = c(time[time <= 0], 0, time[time > 0]) # insert 0
   
-  ## TODO add user supplied functions
-  response = switch(type, # 1 = constant, 2 = sin, 3 = sigmoid
-                    rep(reference, length(time)), # constant
-                    reference*sin(pi/2 + pi*time/emax), # part of sin curve
-                    reference*(1 - 1/(1 + exp(-time/emax*5))) # 1 - sigmoid
-  ) + rnorm(n_obs, mean = noise[1], sd = noise[2])
-  cbind(id, time, response, true_group)
+  base = switch(type, # 1 = constant, 2 = sin, 3 = sigmoid
+                    100*rep(0, length(time)), # constant
+                    100*sin(2*pi/3 + pi*time/emax), # part of sin curve
+                    100*(1 - 1/(1 + exp(-time/emax*5))) # 1 - sigmoid
+  )
+  base = intercept + (base - base[1])
+  response = base[-1] + rnorm(n_obs, mean = noise[1], sd = noise[2])
+  time = time[-1]
+  cbind(id, time, response, true_group) # removed smin time used to align
 }
 
 #' gendata
@@ -47,13 +50,15 @@ oneid = function(id, n_obs, type, start, end, smin, emax, reference, noise) {
 #' See parameters of {\code{\link{gen_traj_data}}}.
 #' @param m_obs
 #' See parameters of {\code{\link{gen_traj_data}}}.
+#' @param types
+#' See parameters of {\code{\link{gen_traj_data}}}.
+#' @param intercepts
+#' See parameters of {\code{\link{gen_traj_data}}}.
 #' @param s_range
 #' See parameters of {\code{\link{gen_traj_data}}}.
 #' @param e_range
 #' See parameters of {\code{\link{gen_traj_data}}}.
 #' @param min_obs
-#' See parameters of {\code{\link{gen_traj_data}}}.
-#' @param reference
 #' See parameters of {\code{\link{gen_traj_data}}}.
 #' @param noise
 #' See parameters of {\code{\link{gen_traj_data}}}.
@@ -64,18 +69,20 @@ oneid = function(id, n_obs, type, start, end, smin, emax, reference, noise) {
 #' @return 
 #' A list of length `sum(n_id)`, where each element is a matrix output by {\code{\link{oneid}}}.
 #' 
-gendata = function(n_id, m_obs, s_range, e_range, min_obs, reference, noise) {
+gendata = function(n_id, types, intercepts, m_obs, s_range, e_range, min_obs,
+                   noise) {
   idr = c(1, 3*sum(n_id)) # id range to pick (so non-consecutive)
   t_id = sum(n_id)
   
   start = round(runif(t_id, min = s_range[1], max = s_range[2]))
   end = floor(runif(t_id, min = e_range[1], max = e_range[2]))
   n_obs = min_obs + rpois(t_id, lambda = m_obs)
-  type = sample(1:length(n_id), t_id, replace = TRUE, prob = n_id/t_id)
+  type = sample(1:length(types), t_id, replace = TRUE, prob = n_id/t_id)
   id_vec = sample.int(idr[2], t_id, replace = FALSE) # non-consecutive
   lapply(1:t_id,   # construct by id
-         function(i) oneid(id_vec[i], n_obs[i], type[i], start[i], end[i],
-                           smin = min(start), emax = max(end), reference, noise))
+         function(i) oneid(id_vec[i], n_obs[i], types[type[i]], intercepts[type[i]],
+                           start[i], end[i], smin = min(start), emax = max(end),
+                           noise))
 }
 
 #' Data Generators
@@ -93,12 +100,20 @@ gendata = function(n_id, m_obs, s_range, e_range, min_obs, reference, noise) {
 #' distributed in e_range. Number of observations in a trajectory is 
 #' Poisson(m_obs). The result is a number of trajectories, all starting at 
 #' time 0, with different time spans, and with independently different numbers 
-#' of observations within the time spans. Each trajectory follows a randomly 
-#' selected response function with added N(mean, sd) error.
+#' of observations within the time spans. Each trajectory follows one of three 
+#' possible response functions possibly with a different mean and with added
+#' N(mean, sd) error.
 #' 
 #' @param n_id 
 #' Vector whose length is the number of clusters, giving the number of id's to
 #' generate in each cluster.
+#' @param types 
+#' A vector of integers from `c(1, 2, 3)` of same length as n_id, indicating 
+#' curve type: constant, sine portion, sigmoid portion, respectively.
+#' @param intercepts
+#' A vector of first responses at minimum time for the curve base vectors of 
+#' same length as n_id. Each `type`-`intercept` combination should be unique for 
+#' unique clusters.
 #' @param m_obs 
 #' Mean number of observation per id. Provides \code{lambda} parameter in
 #' \code{\link[stats]{rpois}}.
@@ -108,9 +123,6 @@ gendata = function(n_id, m_obs, s_range, e_range, min_obs, reference, noise) {
 #' @param e_range 
 #' A vector of length 2, giving the min and max limits of uniformly generated
 #' end observation time.
-#' @param reference 
-#' A nominal response value (for example, blood pressure is near 100, which
-#' is the default)
 #' @param noise 
 #' Vector of length 2 giving the *mean* and *sd* of added N(mean, sd) noise.
 #' @param min_obs 
@@ -120,18 +132,19 @@ gendata = function(n_id, m_obs, s_range, e_range, min_obs, reference, noise) {
 #'   `id`, `time`, `response`, and `true_group`.
 #'   
 #' @examples
-#' data = gen_traj_data(n_id = c(50, 100), m_obs = 20, s_range = c(-365, -14),
-#'               e_range = c(0.5*365, 2*365))
+#' data = gen_traj_data(n_id = c(50, 100), types = c(1, 2), 
+#'   intercepts = c(100, 80), m_obs = 20, s_range = c(-365, -14), 
+#'   e_range = c(0.5*365, 2*365))
 #' head(data)
 #' tail(data)
 #' 
 #' @importFrom stats dist rnorm rpois runif
 #' @export
-gen_traj_data = function(n_id, m_obs, s_range, e_range, reference = 100,
-                         noise = c(0, abs(reference/20)), min_obs = 3)
+gen_traj_data = function(n_id, types, intercepts, m_obs, s_range, e_range,
+                         noise = c(0, abs(mean(intercepts)/20)), min_obs = 3)
 {
-  if(length(n_id) > 3) stop("Don't know how to generate more than 3 clusters")
-  id_list = gendata(n_id, m_obs, s_range, e_range, min_obs, reference, noise)
+  id_list = gendata(n_id, types, intercepts, m_obs, s_range, e_range, min_obs, 
+                    noise)
   id_mat = do.call(rbind, id_list)
   data.table::data.table(id_mat)
 }