+++ NEWS.md
@@ -1,3 +1,14 @@
+# clustra v0.2.0
+* Reduced vignette dependence to only base graphics
+* New function `plot_sample` built on base graphics
+* New function `plot_smooths` built on base graphics
+* New function `plot_silhouette` built on base graphics
+* Expanded capability of `gen_traj_data` beyond 3 clusters with `type` and `intercept` parameters
+* Modified parameters in `clustra`
+* `start_groups` now has added option "distant"
+* Added parameter `starts` in `clustra_sil` and `clustra_rand`
+* New internal functions `kchoose` and `ic_fun` (not exported)
+
 # clustra v0.1.6
 * Original ids are now included in `clustra` output
 * Added the ability to reuse previous clustra runs by `clustra_sil` for silhouette plots